+++ tests/testthat/test-ts.R
@@ -1,0 +1,8 @@
+context("ts")
+
+test_that("filter and lag throw errors", {
+  x <- ts(1:10)
+
+  expect_error(filter(x), "dplyr::filter")
+  expect_error(lag(x), "dplyr::lag")
+})